Use GUIConsole as superclass of Console for environments with AWT support
